requirements.txt
torch==2.0.0
transformers==4.30.0
datasets==2.14.0
scikit-learn==1.2.2
pandas==1.5.3
numpy==1.23.5
matplotlib==3.7.1
seaborn==0.12.2
tqdm==4.65.0
pyyaml==6.0
wandb==0.13.10
evaluate==0.4.0
